Fix indexing to make dotfile FAs look like ASCII maps.
Right now, the y-axis is messed up on dot graph renderings of the finite
automata that we generate from the ASCII mazes. Basically dot and we
index these in the opposite order, leading to the y axis looking flipped
on the dot visualizations.

This commit reverses our y axis, fixing this problem.